Directory reorganization, some renaming.
