+++ assets/src/service-worker/index.js
